--- src/main/java/com/moblie/management/local/order/repository/OrderProductRepository.java
@@ -1,7 +1,0 @@
-package com.moblie.management.local.order.repository;
-
-import com.moblie.management.local.order.model.OrderProduct;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {
-}
