--- src/main/java/com/moblie/management/local/order/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.moblie.management.local.order.repository;
-
-import com.moblie.management.local.order.model.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
