+++ src/main/java/com/moblie/management/local/order/repository/order/OrderRepository.java
@@ -1,0 +1,12 @@
+package com.moblie.management.local.order.repository.order;
+
+import com.moblie.management.local.order.model.Order;
+import com.moblie.management.local.order.model.OrderProduct;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    Order findByTrackingId(String trackingId);
+}
